requirements.txt
asgiref==3.5.0
Django>=3.2,<3.3
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
django-cors-headers==3.11.0
django-filter==21.1
pytz==2021.3
Pillow==9.0.0
sqlparse==0.4.2
gunicorn==20.1.0
psycopg2==2.9.3
django-dbbackup==3.3.0
python-dateutil==2.8.2
django-cleanup==6.0.0
django-rest-swagger>=2.2.0, <2.3.0
drf-yasg==1.21.5
python-dotenv~=1.0.0
apscheduler==3.10.1
pandas==1.4.4
